9-Go.json: Normalize installCommandHtml
Always use `go get NAME`; don't sometimes include `-u` and sometimes not.
Have NAME be a hyperlink to the godocs.

Exception: Because of the trailing `/v3` for gbrlsnchs/jwt, it MUST be
installed using Go modules, so I prefixed it with `GO111MODULE=on`.  The
`/v3` is not present in the href because godoc.org does not yet suppot
modules.  Unfortunately, this means that the linked docs might disagree
with the version they'd get from installing it.  This can't be helped.