Revert "Use node-jose to parse the jwks_uri from OIDC Discovery"
This reverts a few properties that were introduced there:
 a. When selecting the key from the key set, it no longer checks that the
    'alg' matches or that 'use'='sig'.
 b. It no longer checks that exactly 1 key matches.
 c. It no longer does whatever node-jose was doing that the returned PEM
    didn't verbatim match the x5c.

If any of these properties are desired, they should be re-implemented in
getKeyFromJwkKeySetUrl(), so that the are applied to all JWKS parsing,
instead of just when we got the JWKS from OIDC-Discovery (having different
behavior for that is an unexpected special case).

Reverting property (c) is nice, because in order to accommodate it,
derived PEM was copy/pasted around in the test suite, which lead to an
improper merge in f1d8db521670800f13cc4c57d65e9a52f6b4010a, when one side
of the merge changed the x5c to something with more bits, and on the other
side of the merge, the support code for (c) added "e" and "n" parameters
to the key.  The newly-added "e" and "n" didn't agree with the
newly-changed "x5c", and the test suite has been failing ever since.

This reverts commit 4ebebf25853bbb59b6aa76083dd2aeb284180fdb.