Update 3-Java.json
Hello,

I've just released implementations of JOSE RFCs in an Inverno framework module. The library is quite complete and can actually be used independently from the framework so I thought this might interest some people.

I'm not sure about the process to get it registered on jwt.io or whether this is even possible which is why I've made this PR.

Please let me know if more is needed (e.g. tests or other...).

Regards